ui slice to TS
